Refactor test to use useCases rather than hitting the API
