+++ packages/api/src/routes/v1/cases/case.e2e.test.ts
@@ -1,9 +1,14 @@
 import type { CaseDto } from "@moj-bichard7/common/types/Case"
 import type { FastifyInstance } from "fastify"
 
+import { auditLogEventLookup as AuditLogEventLookup } from "@moj-bichard7/common/types/AuditLogEvent"
+import EventCategory from "@moj-bichard7/common/types/EventCategory"
+import EventCode from "@moj-bichard7/common/types/EventCode"
 import { isError } from "@moj-bichard7/common/types/Result"
 import { randomUUID } from "crypto"
 import { OK } from "http-status"
+
+import type { OutputApiAuditLog } from "../../../types/AuditLog"
 
 import { V1 } from "../../../endpoints/versionedEndpoints"
 import { testAhoJsonStr, testAhoXml } from "../../../tests/helpers/ahoHelper"
@@ -12,6 +17,7 @@ import { mockInputApiAuditLog } from "../../../tests/helpers/mockAuditLogs"
 import { SetupAppEnd2EndHelper } from "../../../tests/helpers/setupAppEnd2EndHelper"
 import { createUserAndJwtToken, createUsers, generateJwtForUser } from "../../../tests/helpers/userHelper"
 import createAuditLog from "../../../useCases/createAuditLog"
+import FetchById from "../../../useCases/fetchAuditLogs/FetchById"
 
 describe("/v1/case e2e", () => {
   const endpoint = V1.Case
@@ -206,5 +212,50 @@ describe("/v1/case e2e", () => {
       const responseJson: CaseDto = (await response.json()) satisfies CaseDto
       expect(responseJson.errorLockedByUsername).toBe(expectedLockedByUsername)
     })
+  })
+
+  it("should create a new audit log event when locking a case", async () => {
+    const messageId = randomUUID()
+
+    const [user] = await createUsers(helper.postgres, 3)
+    const jwtToken = await generateJwtForUser(user)
+
+    const testCase = await createCase(helper.postgres, {
+      error_count: 1,
+      error_locked_by_id: null,
+      error_status: 1,
+      message_id: messageId,
+      org_for_police_filter: "01"
+    })
+
+    const auditLog = mockInputApiAuditLog({ caseId: "1", messageId })
+    const result = await createAuditLog(auditLog, helper.dynamo)
+    expect(isError(result)).toBe(false)
+
+    const caseResult = await fetch(`${helper.address}${endpoint.replace(":caseId", testCase.error_id.toString())}`, {
+      headers: {
+        Authorization: `Bearer ${jwtToken}`
+      },
+      method: "GET"
+    })
+
+    expect(caseResult.status).toBe(OK)
+
+    const auditLogJson = await new FetchById(helper.dynamo, messageId).fetch()
+
+    expect(isError(result)).toBe(false)
+    expect(auditLogJson).toBeDefined()
+
+    const auditLogObj = auditLogJson as OutputApiAuditLog
+
+    expect(auditLogObj.events).toHaveLength(1)
+
+    const auditLogEvent = auditLogObj.events?.[0]
+
+    expect(auditLogEvent).toHaveProperty("eventCode", EventCode.ExceptionsLocked)
+    expect(auditLogEvent).toHaveProperty("category", EventCategory.information)
+    expect(auditLogEvent).toHaveProperty("eventSource", "Bichard New UI")
+    expect(auditLogEvent).toHaveProperty("eventType", AuditLogEventLookup[EventCode.ExceptionsLocked])
+    expect(auditLogEvent).toHaveProperty("user", user.username)
   })
 })
